Refactor blog filters
